move restreamer.stats to common.stats
